backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
psycopg[binary]==3.1.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
redis==5.0.3
celery==5.3.6
httpx==0.27.0
pytest==8.1.1
# scikit-learn==1.4.1.post1
# xgboost==2.0.3
# pandas==2.2.1
# numpy==1.26.4
# google-generativeai==0.4.1
# sentence-transformers==2.5.1
# faiss-cpu==1.8.0
# joblib==1.3.2
python-multipart==0.0.9
